+++ cnccode/controller_classes/CTCNC.inc.php
@@ -1151,4 +1151,13 @@ class CTCNC extends Controller
     {
         return $this->dbeUser->getValue(DBEUser::createRenewalSalesOrdersFlag) == 'Y';
     }
+
+    protected function console_log($output, $with_script_tags = true) {
+        $js_code = 'console.log(' . json_encode($output, JSON_HEX_TAG) . 
+    ');';
+        if ($with_script_tags) {
+            $js_code = '<script>' . $js_code . '</script>';
+        }
+        echo $js_code;
+    }
 }